pool.go to be functions on pool
